feat: add CreateDatabase resource and SQL script
Introduced a new localized string resource `CreateDatabase` in the `SqlScripts.Designer.cs` file to retrieve a SQL command for creating a "Flashcards" database if it does not exist. Added the corresponding entry in the `SqlScripts.resx` file with the SQL command value. Additionally, created a new `CreateDatabase.sql` file to encapsulate the database creation logic in a separate script.