refactor: change StartStudySessionAsync to void
Updated the `StartStudySessionAsync` method in both the `IStudySessionsService` interface and the `StudySessionsService` class to change its return type from `Task` to `void`. This change indicates that the method is no longer asynchronous, simplifying its usage by removing the need to await it. All calls to this method have been adjusted accordingly.