feat: add support for Cloze card type in CardsService
The `CardsService` class has been updated to include support for a new card type, `Cloze`. This change was made in two places within the class: once for adding a card and once for updating a card. In both instances, a new case for `CardType.Cloze` was added to the switch expression that determines the appropriate card strategy to use. This allows the service to handle `Cloze` cards in addition to `Flashcard` and `MultipleChoice` cards.