feat: add Box and NextReviewDate columns to Cards table
Added two new columns to the `Cards` table in the `CreateTables.sql` file. The `Box` column is an integer that defaults to 1 and cannot be null, while the `NextReviewDate` column is a date that defaults to the current date and also cannot be null. These additions enhance the functionality of the Cards table for better review management.