test: add unit tests for GetNextBox method
Added a new using directive for `System.Reflection` and introduced additional test cases for the `GetNextBox` method in the `StudySessionsServiceTests` class. The new test method `GetNextBox_ShouldReturnExpectedBox` utilizes the `[Theory]` and `[InlineData]` attributes to cover multiple input scenarios. A private static method `InvokeGetNextBox` was also added to enable reflection-based invocation of the non-public `GetNextBox` method in the `StudySessionsService` class.