refactor: move mapping from services to mapper class
A new method `ToCardDTOList` was added to the `Mapper` class in `Mapper.cs` to convert an `IEnumerable<BaseCard>` to a `List<BaseCardDto>`. This change replaces the previous `foreach` loop and `switch` statement in both the `CardsService` and `StacksService` classes with a call to the new method, simplifying the code and enhancing maintainability. This refactor improves the overall structure and readability of the card conversion logic.