refactor: move study session mapping from service to mapper class
Added a new method `ToStudySessionDTOList` in the `Mapper` class to convert a collection of `StudySession` objects into a list of `StudySessionDto` objects using LINQ.

Updated the `StudySessionsService` class to utilize this new method, replacing a manual loop for conversion. This change simplifies the code and enhances readability by leveraging the new mapping functionality.